Address code review feedback (excluding institution list)
